src/bot/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
gradio>=4.36.0
PyPDF2>=3.0.0
pydantic>=2.6.0
requests>=2.31.0
ruff>=0.12.7 